updating

+++ torchbenchmark/models/timm_efficientnet/__init__.py
@@ -57,7 +57,7 @@ class Model(BenchmarkModel):
 		if precision=='fp16':
 			output = self.model(self.batch.half())
 		elif precision=='bfloat16':
-			output = self.model.bfloat16()(self.batch.bfloat16())
+			output = self.model(self.batch.bfloat16())
 		else:
 			output = self.model(self.batch)
 		nvtx.range_pop()
